Enable "make test" on Travis-CI
This also uses Ubuntu 14.04 "trusty", and installs nmap.  14.04 is required
because the version of nmap included in 12.04 would die with:
    Ncat: Program bug: fd (5) not on list. QUITTING.
if it has multiple connections and the first one died.  This was a problem for
the spiped testing suite, which expected to issue a
    kill $nc_pid
to terminate the ncat server in that test.